+++ CHANGELOG.md
@@ -16,6 +16,10 @@ All notable changes to Jodd project are documented here.
 + **pathref** - methods `on()` and `to()` renamed to `of()` and `get()`.
 + **jerry** - removed `$()` method in favor of `s()` - graal compatible.
 
+### Bug fixes
+
++ **http** - cookies not passed down in `HttpBrowser` after a redirect occurs. 
+
 
 ## [v5.0.13](https://github.com/oblac/jodd/compare/v5.0.12...v5.0.13)